+++ cli/azd/cmd/env.go
@@ -265,6 +265,7 @@ func newEnvNewCmd() *cobra.Command {
 type envNewAction struct {
 	azdCtx             *azdcontext.AzdContext
 	userProfileService *azcli.UserProfileService
+	subResolver        account.SubscriptionTenantResolver
 	accountManager     account.Manager
 	flags              *envNewFlags
 	args               []string
@@ -274,6 +275,7 @@ type envNewAction struct {
 func newEnvNewAction(
 	azdCtx *azdcontext.AzdContext,
 	userProfileService *azcli.UserProfileService,
+	subResolver account.SubscriptionTenantResolver,
 	_ auth.LoggedInGuard,
 	accountManager account.Manager,
 	flags *envNewFlags,
@@ -287,6 +289,7 @@ func newEnvNewAction(
 		flags:              flags,
 		args:               args,
 		console:            console,
+		subResolver:        subResolver,
 	}
 }
 
@@ -302,7 +305,7 @@ func (en *envNewAction) Run(ctx context.Context) (*actions.ActionResult, error) 
 		location:        en.flags.location,
 	}
 	if _, err := createAndInitEnvironment(
-		ctx, &envSpec, en.azdCtx, en.console, en.accountManager, en.userProfileService); err != nil {
+		ctx, &envSpec, en.azdCtx, en.console, en.accountManager, en.userProfileService, en.subResolver); err != nil {
 		return nil, fmt.Errorf("creating new environment: %w", err)
 	}
 
@@ -340,6 +343,7 @@ func newEnvRefreshCmd() *cobra.Command {
 type envRefreshAction struct {
 	azdCtx              *azdcontext.AzdContext
 	projectConfig       *project.ProjectConfig
+	projectManager      project.ProjectManager
 	accountManager      account.Manager
 	azCli               azcli.AzCli
 	env                 *environment.Environment
@@ -356,6 +360,7 @@ func newEnvRefreshAction(
 	projectConfig *project.ProjectConfig,
 	azCli azcli.AzCli,
 	accountManager account.Manager,
+	projectManager project.ProjectManager,
 	env *environment.Environment,
 	commandRunner exec.CommandRunner,
 	flags *envRefreshFlags,
@@ -368,6 +373,7 @@ func newEnvRefreshAction(
 		azdCtx:              azdCtx,
 		azCli:               azCli,
 		accountManager:      accountManager,
+		projectManager:      projectManager,
 		env:                 env,
 		flags:               flags,
 		console:             console,
@@ -416,7 +422,7 @@ func (ef *envRefreshAction) Run(ctx context.Context) (*actions.ActionResult, err
 		}
 	}
 
-	if err = ef.projectConfig.Initialize(ctx, ef.env, ef.commandRunner); err != nil {
+	if err = ef.projectManager.Initialize(ctx, ef.projectConfig); err != nil {
 		return nil, err
 	}