+++ cli/azd/cmd/infra_create.go
@@ -77,6 +77,8 @@ Depending on what Azure resources are created, running this command might take a
 type infraCreateAction struct {
 	flags               *infraCreateFlags
 	accountManager      account.Manager
+	projectManager      project.ProjectManager
+	resourceManager     project.ResourceManager
 	azdCtx              *azdcontext.AzdContext
 	azCli               azcli.AzCli
 	env                 *environment.Environment
@@ -91,6 +93,8 @@ type infraCreateAction struct {
 func newInfraCreateAction(
 	flags *infraCreateFlags,
 	accountManager account.Manager,
+	projectManager project.ProjectManager,
+	resourceManager project.ResourceManager,
 	azdCtx *azdcontext.AzdContext,
 	projectConfig *project.ProjectConfig,
 	azCli azcli.AzCli,
@@ -104,6 +108,8 @@ func newInfraCreateAction(
 	return &infraCreateAction{
 		flags:               flags,
 		accountManager:      accountManager,
+		projectManager:      projectManager,
+		resourceManager:     resourceManager,
 		azdCtx:              azdCtx,
 		azCli:               azCli,
 		env:                 env,
@@ -123,7 +129,7 @@ func (i *infraCreateAction) Run(ctx context.Context) (*actions.ActionResult, err
 		TitleNote: "Provisioning Azure resources can take some time"},
 	)
 
-	if err := i.projectConfig.Initialize(ctx, i.env, i.commandRunner); err != nil {
+	if err := i.projectManager.Initialize(ctx, i.projectConfig); err != nil {
 		return nil, err
 	}
 
@@ -210,7 +216,7 @@ func (i *infraCreateAction) Run(ctx context.Context) (*actions.ActionResult, err
 	return &actions.ActionResult{
 		Message: &actions.ResultMessage{
 			Header:   "Your project has been provisioned!",
-			FollowUp: getResourceGroupFollowUp(ctx, i.formatter, i.azCli, i.projectConfig, i.env),
+			FollowUp: getResourceGroupFollowUp(ctx, i.formatter, i.azCli, i.projectConfig, i.resourceManager, i.env),
 		},
 	}, nil
 }