+++ cli/azd/internal/repository/initializer.go
@@ -40,7 +40,8 @@ func NewInitializer(
 // Initializes a local repository in the project directory from a remote repository.
 //
 // A confirmation prompt is displayed for any existing files to be overwritten.
-func (i *Initializer) Initialize(ctx context.Context,
+func (i *Initializer) Initialize(
+	ctx context.Context,
 	azdCtx *azdcontext.AzdContext,
 	templateUrl string,
 	templateBranch string) error {
@@ -68,7 +69,7 @@ func (i *Initializer) Initialize(ctx context.Context,
 		return err
 	}
 
-	err = i.promptForDuplicates(ctx, staging, target)
+	skipStagingFiles, err := i.promptForDuplicates(ctx, staging, target)
 	if err != nil {
 		return err
 	}
@@ -78,8 +79,19 @@ func (i *Initializer) Initialize(ctx context.Context,
 		return err
 	}
 
-	if err := copy.Copy(staging, target); err != nil {
-		return fmt.Errorf("copying template contents: %w", err)
+	options := copy.Options{}
+	if skipStagingFiles != nil {
+		options.Skip = func(fileInfo os.FileInfo, src, dest string) (bool, error) {
+			if _, shouldSkip := skipStagingFiles[src]; shouldSkip {
+				return true, nil
+			}
+
+			return false, nil
+		}
+	}
+
+	if err := copy.Copy(staging, target, options); err != nil {
+		return fmt.Errorf("copying template contents from temp staging directory: %w", err)
 	}
 
 	err = i.writeAzdAssets(ctx, azdCtx)
@@ -124,7 +136,10 @@ func (i *Initializer) fetchCode(
 	return executableFilePaths, nil
 }
 
-func (i *Initializer) promptForDuplicates(ctx context.Context, staging string, target string) error {
+// promptForDuplicates prompts the user for any duplicate files detected.
+// The list of absolute source file paths to skip are returned.
+func (i *Initializer) promptForDuplicates(
+	ctx context.Context, staging string, target string) (skipSourceFiles map[string]struct{}, err error) {
 	log.Printf(
 		"template init, checking for duplicates. source: %s target: %s",
 		staging,
@@ -133,34 +148,46 @@ func (i *Initializer) promptForDuplicates(ctx context.Context, staging string, t
 
 	duplicateFiles, err := determineDuplicates(staging, target)
 	if err != nil {
-		return fmt.Errorf("checking for overwrites: %w", err)
+		return nil, fmt.Errorf("checking for overwrites: %w", err)
 	}
 
 	if len(duplicateFiles) > 0 {
 		i.console.StopSpinner(ctx, "", input.StepDone)
 		i.console.MessageUxItem(ctx, &ux.WarningMessage{
-			Description: "the following files will be overwritten with the versions from the template:",
+			Description: "The following files are present both locally and in the template:",
 		})
 
 		for _, file := range duplicateFiles {
 			i.console.Message(ctx, fmt.Sprintf(" * %s", file))
 		}
 
-		overwrite, err := i.console.Confirm(ctx, input.ConsoleOptions{
-			Message:      "Overwrite files with versions from template?",
-			DefaultValue: false,
+		selection, err := i.console.Select(ctx, input.ConsoleOptions{
+			Message: "What would you like to do with these files?",
+			Options: []string{
+				"Overwrite with versions from template",
+				"Keep my existing files unchanged",
+			},
 		})
 
 		if err != nil {
-			return fmt.Errorf("prompting to overwrite: %w", err)
+			return nil, fmt.Errorf("prompting to overwrite: %w", err)
 		}
 
-		if !overwrite {
-			return errors.New("confirmation declined")
+		switch selection {
+		case 0: // overwrite
+			return nil, nil
+		case 1: // keep
+			skipSourceFiles = make(map[string]struct{}, len(duplicateFiles))
+			for _, file := range duplicateFiles {
+				// this also cleans the result, which is important for matching
+				sourceFile := filepath.Join(staging, file)
+				skipSourceFiles[sourceFile] = struct{}{}
+			}
+			return skipSourceFiles, nil
 		}
 	}
 
-	return nil
+	return nil, nil
 }
 
 func (i *Initializer) gitInitialize(ctx context.Context,
@@ -227,7 +254,7 @@ func parseExecutableFiles(stagedFilesOutput string) ([]string, error) {
 			// Advance to past '\t', taking the remainder which is <file>
 			_, filepath, found := strings.Cut(scanner.Text()[advance:], "\t")
 			if !found {
-				return nil, errors.New("invalid staged files output format. Missing file path.")
+				return nil, errors.New("invalid staged files output format, missing file path")
 			}
 
 			executableFiles = append(executableFiles, filepath)
@@ -269,7 +296,7 @@ func (i *Initializer) InitializeEmpty(ctx context.Context, azdCtx *azdcontext.Az
 func (i *Initializer) writeAzdAssets(ctx context.Context, azdCtx *azdcontext.AzdContext) error {
 	// Check to see if `azure.yaml` exists, and if it doesn't, create it.
 	if _, err := os.Stat(azdCtx.ProjectPath()); errors.Is(err, os.ErrNotExist) {
-		_, err = project.NewProject(azdCtx.ProjectPath(), azdCtx.GetDefaultProjectName())
+		_, err = project.New(ctx, azdCtx.ProjectPath(), azdCtx.GetDefaultProjectName())
 		i.console.MessageUxItem(ctx,
 			&ux.DoneMessage{Message: fmt.Sprintf("Created a new %s file", azdcontext.ProjectFileName)})
 
@@ -360,7 +387,7 @@ func (i *Initializer) writeAzdAssets(ctx context.Context, azdCtx *azdcontext.Azd
 }
 
 // Returns files that are both present in source and target.
-// The returned files are full paths to the target files.
+// The files returned are expressed in their relative paths to source/target.
 func determineDuplicates(source string, target string) ([]string, error) {
 	var duplicateFiles []string
 	if err := filepath.WalkDir(source, func(path string, d fs.DirEntry, walkErr error) error {