+++ cli/azd/pkg/exec/util_test.go
@@ -87,10 +87,11 @@ func TestAppendEnv(t *testing.T) {
 	require.Equal(t, expectedEnv, actualEnv)
 }
 
-func TestRunCommandList(t *testing.T) {
-	res, err := RunCommandList(context.Background(), []string{
+func TestRunList(t *testing.T) {
+	runner := NewCommandRunner(os.Stdin, os.Stdout, os.Stderr)
+	res, err := runner.RunList(context.Background(), []string{
 		"git --version",
-	}, nil, "")
+	}, RunArgs{})
 
 	if err != nil {
 		t.Errorf("failed to run command list: %v", err)
@@ -107,7 +108,6 @@ func TestRunCommandList(t *testing.T) {
 	if !regexp.MustCompile(`git version\s+\d+\.\d+\.\d+`).Match([]byte(res.Stdout)) {
 		t.Errorf("stdout did not contain 'git version' output")
 	}
-
 }
 
 func TestRunCapturingStderr(t *testing.T) {
@@ -195,9 +195,9 @@ func TestRedactSensitiveData(t *testing.T) {
 
 		{scenario: "SWADeploymentToken",
 			// nolint:lll
-			input: `npx -y @azure/static-web-apps-cli@1.0.0 deploy --tenant-id abc-123 --subscription-id abc-123 --resource-group r --app-name app-name --app-location / --output-location . --env default --no-use-keychain --deployment-token abc-123`,
+			input: `npx -y @azure/static-web-apps-cli@1.0.6 deploy --tenant-id abc-123 --subscription-id abc-123 --resource-group r --app-name app-name --app-location / --output-location . --env default --no-use-keychain --deployment-token abc-123`,
 			// nolint:lll
-			expected: `npx -y @azure/static-web-apps-cli@1.0.0 deploy --tenant-id abc-123 --subscription-id abc-123 --resource-group r --app-name app-name --app-location / --output-location . --env default --no-use-keychain --deployment-token <redacted>`},
+			expected: `npx -y @azure/static-web-apps-cli@1.0.6 deploy --tenant-id abc-123 --subscription-id abc-123 --resource-group r --app-name app-name --app-location / --output-location . --env default --no-use-keychain --deployment-token <redacted>`},
 
 		{scenario: "DockerLoginUsernameAndPassword",
 			input:    `docker login --username crusername123 --password abc123 some.azurecr.io`,