+++ cli/azd/pkg/project/project_config.go
@@ -2,25 +2,9 @@ package project
 
 import (
 	"context"
-	"fmt"
-	"log"
-	"os"
-	"path/filepath"
-	"sort"
 
-	"github.com/azure/azure-dev/cli/azd/internal/telemetry"
-	"github.com/azure/azure-dev/cli/azd/internal/telemetry/fields"
-	"github.com/azure/azure-dev/cli/azd/pkg/account"
-	"github.com/azure/azure-dev/cli/azd/pkg/alphafeatures"
-	"github.com/azure/azure-dev/cli/azd/pkg/environment"
-	"github.com/azure/azure-dev/cli/azd/pkg/exec"
 	"github.com/azure/azure-dev/cli/azd/pkg/ext"
 	"github.com/azure/azure-dev/cli/azd/pkg/infra/provisioning"
-	"github.com/azure/azure-dev/cli/azd/pkg/input"
-	"github.com/azure/azure-dev/cli/azd/pkg/osutil"
-	"github.com/azure/azure-dev/cli/azd/pkg/tools"
-	"github.com/azure/azure-dev/cli/azd/pkg/tools/azcli"
-	"gopkg.in/yaml.v3"
 )
 
 // ProjectConfig is the top level object serialized into an azure.yaml file.
@@ -43,28 +27,6 @@ type ProjectConfig struct {
 type PipelineOptions struct {
 	Provider string `yaml:"provider"`
 }
-
-const (
-	ProjectEventDeploy     ext.Event = "deploy"
-	ProjectEventProvision  ext.Event = "provision"
-	ServiceEventEnvUpdated ext.Event = "environment updated"
-	ServiceEventRestore    ext.Event = "restore"
-	ServiceEventPackage    ext.Event = "package"
-	ServiceEventDeploy     ext.Event = "deploy"
-)
-
-var (
-	ProjectEvents []ext.Event = []ext.Event{
-		ProjectEventProvision,
-		ProjectEventDeploy,
-	}
-	ServiceEvents []ext.Event = []ext.Event{
-		ServiceEventEnvUpdated,
-		ServiceEventRestore,
-		ServiceEventPackage,
-		ServiceEventDeploy,
-	}
-)
 
 // Project lifecycle event arguments
 type ProjectLifecycleEventArgs struct {
@@ -91,153 +53,4 @@ func (p *ProjectConfig) HasService(name string) bool {
 	}
 
 	return false
-}
-
-// GetProject constructs a Project from the project configuration
-// This also performs project validation
-func (pc *ProjectConfig) GetProject(
-	ctx context.Context,
-	env *environment.Environment,
-	console input.Console,
-	azCli azcli.AzCli,
-	commandRunner exec.CommandRunner,
-	accountManager account.Manager,
-	alphaFeatureManager *alphafeatures.AlphaFeatureManager,
-) (*Project, error) {
-	serviceMap := map[string]*Service{}
-
-	project := Project{
-		Name:     pc.Name,
-		Metadata: pc.Metadata,
-		Config:   pc,
-		Services: make([]*Service, 0),
-	}
-
-	resourceGroupName, err := GetResourceGroupName(ctx, azCli, pc, env)
-	if err != nil {
-		return nil, err
-	}
-	project.ResourceGroupName = resourceGroupName
-
-	for key, serviceConfig := range pc.Services {
-		service, err := serviceConfig.GetService(
-			ctx, &project, env, azCli, accountManager, commandRunner, console, alphaFeatureManager)
-
-		if err != nil {
-			return nil, fmt.Errorf("creating service %s: %w", key, err)
-		}
-
-		serviceMap[key] = service
-	}
-
-	// Sort services by friendly name an then collect them into a list. This provides a stable ordering of services.
-	serviceKeys := make([]string, 0, len(serviceMap))
-	for k := range serviceMap {
-		serviceKeys = append(serviceKeys, k)
-	}
-	sort.Strings(serviceKeys)
-
-	for _, key := range serviceKeys {
-		project.Services = append(project.Services, serviceMap[key])
-	}
-
-	return &project, nil
-}
-
-// Saves the current instance back to the azure.yaml file
-func (p *ProjectConfig) Save(projectPath string) error {
-	projectBytes, err := yaml.Marshal(p)
-	if err != nil {
-		return fmt.Errorf("marshalling project yaml: %w", err)
-	}
-
-	err = os.WriteFile(projectPath, projectBytes, osutil.PermissionFile)
-	if err != nil {
-		return fmt.Errorf("saving project file: %w", err)
-	}
-
-	p.Path = projectPath
-
-	return nil
-}
-
-// ParseProjectConfig will parse a project from a yaml string and return the project configuration
-func ParseProjectConfig(yamlContent string) (*ProjectConfig, error) {
-	var projectFile ProjectConfig
-
-	if err := yaml.Unmarshal([]byte(yamlContent), &projectFile); err != nil {
-		return nil, fmt.Errorf(
-			"unable to parse azure.yaml file. Please check the format of the file, "+
-				"and also verify you have the latest version of the CLI: %w",
-			err,
-		)
-	}
-
-	projectFile.EventDispatcher = ext.NewEventDispatcher[ProjectLifecycleEventArgs](ProjectEvents...)
-
-	for key, svc := range projectFile.Services {
-		svc.Name = key
-		svc.Project = &projectFile
-		svc.EventDispatcher = ext.NewEventDispatcher[ServiceLifecycleEventArgs](ServiceEvents...)
-
-		// By convention, the name of the infrastructure module to use when doing an IaC based deployment is the friendly
-		// name of the service. This may be overridden by the `module` property of `azure.yaml`
-		if svc.Module == "" {
-			svc.Module = key
-		}
-
-		if svc.Language == "" || svc.Language == "csharp" || svc.Language == "fsharp" {
-			svc.Language = "dotnet"
-		}
-	}
-
-	return &projectFile, nil
-}
-
-func (p *ProjectConfig) Initialize(
-	ctx context.Context, env *environment.Environment, commandRunner exec.CommandRunner,
-) error {
-	var allTools []tools.ExternalTool
-	for _, svc := range p.Services {
-		frameworkService, err := svc.GetFrameworkService(ctx, env, commandRunner)
-		if err != nil {
-			return fmt.Errorf("getting framework services: %w", err)
-		}
-		if err := frameworkService.Initialize(ctx); err != nil {
-			return err
-		}
-
-		requiredTools := frameworkService.RequiredExternalTools()
-		allTools = append(allTools, requiredTools...)
-	}
-
-	if err := tools.EnsureInstalled(ctx, tools.Unique(allTools)...); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// LoadProjectConfig loads the azure.yaml configuring into an viewable structure
-// This does not evaluate any tooling
-func LoadProjectConfig(projectPath string) (*ProjectConfig, error) {
-	log.Printf("Reading project from file '%s'\n", projectPath)
-	bytes, err := os.ReadFile(projectPath)
-	if err != nil {
-		return nil, fmt.Errorf("reading project file: %w", err)
-	}
-
-	yaml := string(bytes)
-
-	projectConfig, err := ParseProjectConfig(yaml)
-	if err != nil {
-		return nil, fmt.Errorf("parsing project file: %w", err)
-	}
-
-	if projectConfig.Metadata != nil {
-		telemetry.SetUsageAttributes(fields.StringHashed(fields.TemplateIdKey, projectConfig.Metadata.Template))
-	}
-
-	projectConfig.Path = filepath.Dir(projectPath)
-	return projectConfig, nil
 }