+++ cli/azd/pkg/project/service_target_aks_test.go
@@ -11,11 +11,12 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/containerregistry/armcontainerregistry"
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/containerservice/armcontainerservice/v2"
+	"github.com/azure/azure-dev/cli/azd/pkg/async"
 	"github.com/azure/azure-dev/cli/azd/pkg/convert"
 	"github.com/azure/azure-dev/cli/azd/pkg/environment"
-	"github.com/azure/azure-dev/cli/azd/pkg/environment/azdcontext"
 	"github.com/azure/azure-dev/cli/azd/pkg/exec"
 	"github.com/azure/azure-dev/cli/azd/pkg/infra"
 	"github.com/azure/azure-dev/cli/azd/pkg/osutil"
@@ -23,20 +24,19 @@ import (
 	"github.com/azure/azure-dev/cli/azd/pkg/tools/docker"
 	"github.com/azure/azure-dev/cli/azd/pkg/tools/kubectl"
 	"github.com/azure/azure-dev/cli/azd/test/mocks"
+	"github.com/azure/azure-dev/cli/azd/test/mocks/mockaccount"
 	"github.com/azure/azure-dev/cli/azd/test/ostest"
-	"github.com/benbjohnson/clock"
 	"github.com/stretchr/testify/require"
 	"gopkg.in/yaml.v3"
 )
 
 func Test_NewAksTarget(t *testing.T) {
 	mockContext := mocks.NewMockContext(context.Background())
-	serviceConfig := createServiceConfig("")
+	serviceConfig := createTestServiceConfig("./src/api", AksTarget, ServiceLanguageTypeScript)
 	env := createEnv()
 
-	serviceTarget, err := createServiceTarget(mockContext, serviceConfig, env)
+	serviceTarget := createServiceTarget(mockContext, serviceConfig, env)
 
-	require.NoError(t, err)
 	require.NotNil(t, serviceTarget)
 	require.NotNil(t, serviceConfig)
 }
@@ -46,108 +46,106 @@ func Test_Required_Tools(t *testing.T) {
 	ostest.Chdir(t, tempDir)
 
 	mockContext := mocks.NewMockContext(context.Background())
-	err := setupMocks(mockContext)
+	err := setupMocksForAksTarget(mockContext)
 	require.NoError(t, err)
 
-	serviceConfig := createServiceConfig(tempDir)
+	serviceConfig := createTestServiceConfig(tempDir, AksTarget, ServiceLanguageTypeScript)
 	env := createEnv()
 
-	serviceTarget, err := createServiceTarget(mockContext, serviceConfig, env)
-	require.NoError(t, err)
+	serviceTarget := createServiceTarget(mockContext, serviceConfig, env)
 
-	requiredTools := serviceTarget.RequiredExternalTools()
+	requiredTools := serviceTarget.RequiredExternalTools(*mockContext.Context)
 	require.Len(t, requiredTools, 2)
 	require.Implements(t, new(docker.Docker), requiredTools[0])
 	require.Implements(t, new(kubectl.KubectlCli), requiredTools[1])
 }
 
-func Test_Deploy_HappyPath(t *testing.T) {
+func Test_Package_Publish_HappyPath(t *testing.T) {
 	tempDir := t.TempDir()
 	ostest.Chdir(t, tempDir)
 
 	mockContext := mocks.NewMockContext(context.Background())
-	err := setupMocks(mockContext)
+	err := setupMocksForAksTarget(mockContext)
 	require.NoError(t, err)
 
-	serviceConfig := createServiceConfig(tempDir)
+	serviceConfig := createTestServiceConfig(tempDir, AksTarget, ServiceLanguageTypeScript)
 	env := createEnv()
 
-	serviceTarget, err := createServiceTarget(mockContext, serviceConfig, env)
-	require.NoError(t, err)
-
+	serviceTarget := createServiceTarget(mockContext, serviceConfig, env)
 	err = setupK8sManifests(t, serviceConfig)
 	require.NoError(t, err)
 
-	azdContext := azdcontext.NewAzdContextWithDirectory(tempDir)
-	progressChan := createAndLogProgress()
+	packageTask := serviceTarget.Package(
+		*mockContext.Context,
+		serviceConfig,
+		&ServicePackageResult{
+			PackagePath: "IMAGE_ID",
+			Details: &dockerPackageResult{
+				ImageTag:    "IMAGE_TAG",
+				LoginServer: "REGISTRY.azurecr.io",
+			},
+		},
+	)
+	logProgress(packageTask)
+	packageResult, err := packageTask.Await()
 
-	result, err := serviceTarget.Deploy(*mockContext.Context, azdContext, "", progressChan)
 	require.NoError(t, err)
-	require.NotNil(t, result)
-	require.Equal(t, AksTarget, result.Kind)
-	require.NotNil(t, env.Values["SERVICE_SVC_IMAGE_NAME"])
-	require.IsType(t, new(kubectl.Deployment), result.Details)
-	require.Greater(t, len(result.Endpoints), 0)
+	require.NotNil(t, packageResult)
+	require.IsType(t, new(dockerPackageResult), packageResult.Details)
+
+	scope := environment.NewTargetResource("SUB_ID", "RG_ID", "CLUSTER_NAME", string(infra.AzureResourceTypeManagedCluster))
+	publishTask := serviceTarget.Publish(*mockContext.Context, serviceConfig, packageResult, scope)
+	logProgress(publishTask)
+	publishResult, err := publishTask.Await()
+
+	require.NoError(t, err)
+	require.NotNil(t, publishResult)
+	require.Equal(t, AksTarget, publishResult.Kind)
+	require.IsType(t, new(kubectl.Deployment), publishResult.Details)
+	require.Greater(t, len(publishResult.Endpoints), 0)
+	// New env variable is created
+	require.Equal(t, "IMAGE_TAG", env.Values["SERVICE_API_IMAGE_NAME"])
 }
 
-func Test_Deploy_No_Cluster_Name(t *testing.T) {
+func Test_Publish_No_Cluster_Name(t *testing.T) {
 	tempDir := t.TempDir()
 	ostest.Chdir(t, tempDir)
 
 	mockContext := mocks.NewMockContext(context.Background())
-	err := setupMocks(mockContext)
+	err := setupMocksForAksTarget(mockContext)
 	require.NoError(t, err)
 
-	serviceConfig := createServiceConfig(tempDir)
+	serviceConfig := createTestServiceConfig(tempDir, AksTarget, ServiceLanguageTypeScript)
 	env := createEnv()
 
 	// Simulate AKS cluster name not found in env file
 	delete(env.Values, environment.AksClusterEnvVarName)
 
-	serviceTarget, err := createServiceTarget(mockContext, serviceConfig, env)
-	require.NoError(t, err)
+	serviceTarget := createServiceTarget(mockContext, serviceConfig, env)
+	scope := environment.NewTargetResource("SUB_ID", "RG_ID", "CLUSTER_NAME", string(infra.AzureResourceTypeManagedCluster))
+	packageOutput := &ServicePackageResult{
+		Build: &ServiceBuildResult{BuildOutputPath: "IMAGE_ID"},
+		Details: &dockerPackageResult{
+			ImageTag:    "IMAGE_TAG",
+			LoginServer: env.Values[environment.ContainerRegistryEndpointEnvVarName],
+		},
+	}
 
-	azdContext := azdcontext.NewAzdContextWithDirectory(tempDir)
-	progressChan := createAndLogProgress()
+	publishTask := serviceTarget.Publish(*mockContext.Context, serviceConfig, packageOutput, scope)
+	logProgress(publishTask)
 
-	result, err := serviceTarget.Deploy(*mockContext.Context, azdContext, "", progressChan)
+	publishResult, err := publishTask.Await()
 	require.Error(t, err)
 	require.ErrorContains(t, err, "could not determine AKS cluster")
-	require.Equal(t, ServiceDeploymentResult{}, result)
+	require.Nil(t, publishResult)
 }
 
-func Test_Deploy_No_Container_Registry(t *testing.T) {
+func Test_Publish_No_Admin_Credentials(t *testing.T) {
 	tempDir := t.TempDir()
 	ostest.Chdir(t, tempDir)
 
 	mockContext := mocks.NewMockContext(context.Background())
-	err := setupMocks(mockContext)
-	require.NoError(t, err)
-
-	serviceConfig := createServiceConfig(tempDir)
-	env := createEnv()
-
-	// Simulate container registry endpoint not found in env file
-	delete(env.Values, environment.ContainerRegistryEndpointEnvVarName)
-
-	serviceTarget, err := createServiceTarget(mockContext, serviceConfig, env)
-	require.NoError(t, err)
-
-	azdContext := azdcontext.NewAzdContextWithDirectory(tempDir)
-	progressChan := createAndLogProgress()
-
-	result, err := serviceTarget.Deploy(*mockContext.Context, azdContext, "", progressChan)
-	require.Error(t, err)
-	require.ErrorContains(t, err, "could not determine container registry endpoint")
-	require.Equal(t, ServiceDeploymentResult{}, result)
-}
-
-func Test_Deploy_No_Admin_Credentials(t *testing.T) {
-	tempDir := t.TempDir()
-	ostest.Chdir(t, tempDir)
-
-	mockContext := mocks.NewMockContext(context.Background())
-	err := setupMocks(mockContext)
+	err := setupMocksForAksTarget(mockContext)
 	require.NoError(t, err)
 
 	// Simulate list credentials fail.
@@ -155,19 +153,26 @@ func Test_Deploy_No_Admin_Credentials(t *testing.T) {
 	err = setupListClusterAdminCredentialsMock(mockContext, http.StatusUnauthorized)
 	require.NoError(t, err)
 
-	serviceConfig := createServiceConfig(tempDir)
+	serviceConfig := createTestServiceConfig(tempDir, AksTarget, ServiceLanguageTypeScript)
 	env := createEnv()
 
-	serviceTarget, err := createServiceTarget(mockContext, serviceConfig, env)
-	require.NoError(t, err)
+	serviceTarget := createServiceTarget(mockContext, serviceConfig, env)
+	scope := environment.NewTargetResource("SUB_ID", "RG_ID", "CLUSTER_NAME", string(infra.AzureResourceTypeManagedCluster))
+	packageOutput := &ServicePackageResult{
+		Build: &ServiceBuildResult{BuildOutputPath: "IMAGE_ID"},
+		Details: &dockerPackageResult{
+			ImageTag:    "IMAGE_TAG",
+			LoginServer: env.Values[environment.ContainerRegistryEndpointEnvVarName],
+		},
+	}
 
-	azdContext := azdcontext.NewAzdContextWithDirectory(tempDir)
-	progressChan := createAndLogProgress()
+	publishTask := serviceTarget.Publish(*mockContext.Context, serviceConfig, packageOutput, scope)
+	logProgress(publishTask)
+	publishResult, err := publishTask.Await()
 
-	result, err := serviceTarget.Deploy(*mockContext.Context, azdContext, "", progressChan)
 	require.Error(t, err)
 	require.ErrorContains(t, err, "failed retrieving cluster admin credentials")
-	require.Equal(t, ServiceDeploymentResult{}, result)
+	require.Nil(t, publishResult)
 }
 
 func setupK8sManifests(t *testing.T, serviceConfig *ServiceConfig) error {
@@ -215,7 +220,7 @@ func setupListClusterAdminCredentialsMock(mockContext *mocks.MockContext, status
 	return nil
 }
 
-func setupMocks(mockContext *mocks.MockContext) error {
+func setupMocksForAksTarget(mockContext *mocks.MockContext) error {
 	err := setupListClusterAdminCredentialsMock(mockContext, http.StatusOK)
 	if err != nil {
 		return err
@@ -328,8 +333,8 @@ func setupMocks(mockContext *mocks.MockContext) error {
 				ApiVersion: "apps/v1",
 				Kind:       "Deployment",
 				Metadata: kubectl.ResourceMetadata{
-					Name:      "svc-deployment",
-					Namespace: "svc-namespace",
+					Name:      "api-deployment",
+					Namespace: "api-namespace",
 				},
 			},
 			Spec: kubectl.DeploymentSpec{
@@ -364,8 +369,8 @@ func setupMocks(mockContext *mocks.MockContext) error {
 				ApiVersion: "v1",
 				Kind:       "Service",
 				Metadata: kubectl.ResourceMetadata{
-					Name:      "svc-service",
-					Namespace: "svc-namespace",
+					Name:      "api-service",
+					Namespace: "api-namespace",
 				},
 			},
 			Spec: kubectl.ServiceSpec{
@@ -397,8 +402,8 @@ func setupMocks(mockContext *mocks.MockContext) error {
 				ApiVersion: "networking.k8s.io/v1",
 				Kind:       "Ingress",
 				Metadata: kubectl.ResourceMetadata{
-					Name:      "svc-ingress",
-					Namespace: "svc-namespace",
+					Name:      "api-ingress",
+					Namespace: "api-namespace",
 				},
 			},
 			Spec: kubectl.IngressSpec{
@@ -451,19 +456,6 @@ func createK8sResourceList[T any](resource T) *kubectl.List[T] {
 	}
 }
 
-func createServiceConfig(projectDirectory string) *ServiceConfig {
-	return &ServiceConfig{
-		Project: &ProjectConfig{
-			Name: "project",
-			Path: projectDirectory,
-		},
-		Name:         "svc",
-		RelativePath: "./src",
-		Host:         string(AksTarget),
-		Language:     "js",
-	}
-}
-
 func createEnv() *environment.Environment {
 	return environment.EphemeralWithValues("test", map[string]string{
 		environment.TenantIdEnvVarName:                  "TENANT_ID",
@@ -479,22 +471,23 @@ func createServiceTarget(
 	mockContext *mocks.MockContext,
 	serviceConfig *ServiceConfig,
 	env *environment.Environment,
-) (ServiceTarget, error) {
-	scope := environment.NewTargetResource("SUB_ID", "RG_ID", "CLUSTER_NAME", string(infra.AzureResourceTypeManagedCluster))
+) ServiceTarget {
 	kubeCtl := kubectl.NewKubectl(mockContext.CommandRunner)
 	dockerCli := docker.NewDocker(mockContext.CommandRunner)
-	managedClustersService := azcli.NewManagedClustersService(mockContext.Credentials, mockContext.HttpClient)
-	containerRegistryService := azcli.NewContainerRegistryService(mockContext.Credentials, mockContext.HttpClient, dockerCli)
+	credentialProvider := mockaccount.SubscriptionCredentialProviderFunc(
+		func(_ context.Context, _ string) (azcore.TokenCredential, error) {
+			return mockContext.Credentials, nil
+		})
+
+	managedClustersService := azcli.NewManagedClustersService(credentialProvider, mockContext.HttpClient)
+	containerRegistryService := azcli.NewContainerRegistryService(credentialProvider, mockContext.HttpClient, dockerCli)
 
 	return NewAksTarget(
-		serviceConfig,
 		env,
-		scope,
 		managedClustersService,
 		containerRegistryService,
 		kubeCtl,
 		dockerCli,
-		clock.New(),
 	)
 }
 
@@ -529,14 +522,10 @@ func createTestCluster(clusterName, username string) *kubectl.KubeConfig {
 	}
 }
 
-func createAndLogProgress() chan (string) {
-	progressChan := make(chan (string))
-
+func logProgress[T comparable, P comparable](task *async.TaskWithProgress[T, P]) {
 	go func() {
-		for value := range progressChan {
+		for value := range task.Progress() {
 			log.Println(value)
 		}
 	}()
-
-	return progressChan
 }