+++ cli/azd/test/functional/cli_test.go
@@ -15,6 +15,7 @@ import (
 	"fmt"
 	"io/fs"
 	"log"
+	"net/http"
 	"os"
 	osexec "os/exec"
 	"path"
@@ -25,6 +26,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
 	"github.com/azure/azure-dev/cli/azd/internal/telemetry"
 	"github.com/azure/azure-dev/cli/azd/pkg/environment"
@@ -35,6 +37,7 @@ import (
 	"github.com/azure/azure-dev/cli/azd/pkg/project"
 	"github.com/azure/azure-dev/cli/azd/pkg/tools/azcli"
 	"github.com/azure/azure-dev/cli/azd/test/azdcli"
+	"github.com/azure/azure-dev/cli/azd/test/mocks/mockaccount"
 	"github.com/joho/godotenv"
 	"github.com/sethvargo/go-retry"
 	"github.com/stretchr/testify/assert"
@@ -72,7 +75,7 @@ func Test_CLI_Init_AsksForSubscriptionIdAndCreatesEnvAndProjectFile(t *testing.T
 	require.Regexp(t, regexp.MustCompile(fmt.Sprintf(`AZURE_SUBSCRIPTION_ID="%s"`, testSubscriptionId)+"\n"), string(file))
 	require.Regexp(t, regexp.MustCompile(`AZURE_ENV_NAME="TESTENV"`+"\n"), string(file))
 
-	proj, err := project.LoadProjectConfig(filepath.Join(dir, azdcontext.ProjectFileName))
+	proj, err := project.Load(ctx, filepath.Join(dir, azdcontext.ProjectFileName))
 	require.NoError(t, err)
 
 	require.Equal(t, filepath.Base(dir), proj.Name)
@@ -191,11 +194,16 @@ func Test_CLI_InfraCreateAndDelete(t *testing.T) {
 		t.Fatal("could not create credential")
 	}
 
-	azCli := azcli.NewAzCli(cred, azcli.NewAzCliArgs{})
+	azCli := azcli.NewAzCli(mockaccount.SubscriptionCredentialProviderFunc(
+		func(_ context.Context, _ string) (azcore.TokenCredential, error) {
+			return cred, nil
+		}),
+		http.DefaultClient,
+		azcli.NewAzCliArgs{})
 
 	// Verify that resource groups are created with tag
 	resourceManager := infra.NewAzureResourceManager(azCli)
-	rgs, err := resourceManager.GetResourceGroupsForEnvironment(ctx, env)
+	rgs, err := resourceManager.GetResourceGroupsForEnvironment(ctx, env.GetSubscriptionId(), env.GetEnvName())
 	require.NoError(t, err)
 	require.NotNil(t, rgs)
 
@@ -247,11 +255,16 @@ func Test_CLI_InfraCreateAndDeleteUpperCase(t *testing.T) {
 		t.Fatal("could not create credential")
 	}
 
-	azCli := azcli.NewAzCli(cred, azcli.NewAzCliArgs{})
+	azCli := azcli.NewAzCli(mockaccount.SubscriptionCredentialProviderFunc(
+		func(_ context.Context, _ string) (azcore.TokenCredential, error) {
+			return cred, nil
+		}),
+		http.DefaultClient,
+		azcli.NewAzCliArgs{})
 
 	// Verify that resource groups are created with tag
 	resourceManager := infra.NewAzureResourceManager(azCli)
-	rgs, err := resourceManager.GetResourceGroupsForEnvironment(ctx, env)
+	rgs, err := resourceManager.GetResourceGroupsForEnvironment(ctx, env.GetSubscriptionId(), env.GetEnvName())
 	require.NoError(t, err)
 	require.NotNil(t, rgs)
 
@@ -370,6 +383,7 @@ func Test_CLI_ProjectIsNeeded(t *testing.T) {
 func Test_CLI_NoDebugSpewWhenHelpPassedWithoutDebug(t *testing.T) {
 	stdErrBuf := bytes.Buffer{}
 
+	/* #nosec G204 - Subprocess launched with a potential tainted input or cmd arguments false positive */
 	cmd := osexec.Command(azdcli.GetAzdLocation(), "--help")
 	cmd.Stderr = &stdErrBuf