+++ templates/common/infra/terraform/core/monitor/loganalytics/loganalytics.tf
@@ -1,12 +1,12 @@
 terraform {
   required_providers {
     azurerm = {
-      version = "~>3.18.0"
+      version = "~>3.47.0"
       source  = "hashicorp/azurerm"
     }
     azurecaf = {
       source  = "aztfmod/azurecaf"
-      version = "~>1.2.15"
+      version = "~>1.2.24"
     }
   }
 }